Build experimental features using a build tag
